# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Management System API Documentation",
        "description": "API documentation for the Task Management System",
        "version": "1.0.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get list of tasks",
                "description": "Returns list of tasks",
                "operationId": "getTasksList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "get all tasks"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Task"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "No tasks found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "No tasks found"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "description": "Creates a new task and returns the created task",
                "operationId": "storeTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Task title",
                                        "type": "string",
                                        "example": "Complete project"
                                    },
                                    "description": {
                                        "description": "Task description",
                                        "type": "string",
                                        "example": "Finish the project by the end of the week."
                                    },
                                    "expiration_date": {
                                        "description": "Expiration date",
                                        "type": "string",
                                        "example": "1403/05/31"
                                    },
                                    "priority": {
                                        "description": "Task priority",
                                        "type": "string",
                                        "example": "high"
                                    },
                                    "status": {
                                        "description": "Task status",
                                        "type": "string",
                                        "example": "in_progress"
                                    },
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "get task"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error creating task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "No task found"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/{task_id}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Update an existing task",
                "description": "Updates an existing task and returns the updated task",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "Task title",
                                        "type": "string",
                                        "example": "Complete project"
                                    },
                                    "description": {
                                        "description": "Task description",
                                        "type": "string",
                                        "example": "Finish the project by the end of the week."
                                    },
                                    "expiration_date": {
                                        "description": "Expiration date",
                                        "type": "string",
                                        "example": "1403/05/31"
                                    },
                                    "priority": {
                                        "description": "Task priority",
                                        "type": "string",
                                        "example": "high"
                                    },
                                    "status": {
                                        "description": "Task status",
                                        "type": "string",
                                        "example": "in_progress"
                                    },
                                    "user_id": {
                                        "description": "User ID",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "get task"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Task"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Error updating task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "No task found"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete an existing task",
                "description": "Deletes a task and returns a success response",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "description": "Task ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "task deleted"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Task not found"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Task": {
                "properties": {
                    "id": {
                        "description": "Task ID",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Task title",
                        "type": "string",
                        "example": "Complete project"
                    },
                    "description": {
                        "description": "Task description",
                        "type": "string",
                        "example": "Finish the project by the end of the week."
                    },
                    "status": {
                        "description": "Task status",
                        "type": "string",
                        "example": "in-progress"
                    },
                    "priority": {
                        "description": "Task priority",
                        "type": "string",
                        "example": "high"
                    },
                    "expiration_date": {
                        "description": "Expiration date",
                        "type": "string",
                        "example": "2024-09-10"
                    },
                    "created_at": {
                        "description": "Creation date",
                        "type": "string",
                        "example": "2024-08-20 12:00:00"
                    },
                    "updated_at": {
                        "description": "Last update date",
                        "type": "string",
                        "example": "2024-08-21 14:30:00"
                    }
                },
                "type": "object"
            }
        }
    },
...
}